1.0.50 - TimerSystem agora irá usar TimerMap para mapear a indexar os temporizadores. 	- Tickers foi unido a TimerMap afim de seguir a OO. - TimerListener deverá ser encarregado de saber seu próprio nome. - Novos métodos para trabalhar com tempos em milissegundos, JRagnarokUtil. - Nome dos atributos e definições padrões ao nome das tabelas atualizados, Tables. - Removendo algumas constantes que não serão mais usadas em FileDescriptor. - Atualizando listeners que usavam TimerListener para implementar getName e toString.

+++ JRagnarok Emulador/src/org/diverproject/jragnarok/JRagnarok.java
@@ -40,7 +40,7 @@ public class JRagnarok
 		while (isRunning())
 			if (isOk())
 			{
-				long next = timer.update(timer.tick());
+				int next = timer.update(timer.tick());
 				FileDescriptor.update(next);
 			}
 	}
